Add spread rewriting (BROKEN)
The scope information for temporaries is probably wrong.
It also needs some re-organization.
The following test cases break:

- mjsunit/strong/literals
- mjsunit/regress/regress-osr-in-literal
- mjsunit/es6/spread-array
- mjsunit/regress/regress-deopt-in-array-literal-spread
- mjsunit/harmony/regress/regress-4395

For the first four, it's something related to code generation and
bailout entries:

    # Fatal error in ../../src/full-codegen/full-codegen.cc, line 226
    # # Check failed: bailout_entries_[i].id != id.

The fifth produces wrong results, it's something related to scope
resolution for temporaries.  This does not work as expected either:

    ((x, y = [...x]) => y)([42,17])