Calculate the cost of AST generation
This is just for benchmarking, should not be landed!

This patch introduces a mechanism for calculating the cost of
generating the AST, while parsing, by counting how many nested
parser methods were used for generating a given node.  Each
parser method that *does not* generate a new node but just
propagates an existing one to its caller increases this node's
cost by one.

So far, an ugly listing of the costs of all nodes in the AST
is obtained by running with --trace-parse.